feat: Add entity relationship graph and D3.js dependency
Introduces support for an entity relationship graph by adding a new `entity_relationships` field to the `RegulatorReport` type and updating the `Gemini` service schema. This allows for the visualization of connections between identified entities, enhancing the overall understanding of document data.

Also includes the addition of the `d3` library, which will be used for future graph rendering capabilities.